Pressure sensor temperature now returns a float with temperature in degrees c, rather than deci-degrees c

+++ libraries/AP_SpraySystem/AP_SpraySystem.cpp
@@ -56,7 +56,7 @@ uint32_t AP_SpraySystem::get_current_pressure_mbar()
     return pressure_sensor->get_pressure_mbar();
 }
 
-uint32_t AP_SpraySystem::get_current_temperature_c()
+float AP_SpraySystem::get_current_temperature_c()
 {
     return pressure_sensor->get_temperature_c();
 }
